+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v3.1.0
+
+...
+
 ## v3.0.0
 
 ### Important Notes